backend/requirements.txt
asgiref==3.10.0
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.10.1
sqlparse==0.5.3
django-allauth>=0.57.0
dj-rest-auth[with_social]>=5.0.0
requests>=2.31.0
google-auth>=2.23.0
python-dotenv>=1.0.0
psycopg2-binary==2.9.11
